requirements.txt
beautifulsoup4==4.14.3
blinker==1.9.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.1.8
Flask==3.1.2
idna==3.11
importlib_metadata==8.7.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
requests==2.32.5
soupsieve==2.8.3
typing_extensions==4.15.0
urllib3==2.6.3
Werkzeug==3.1.5
zipp==3.23.0
feedparser
duckduckgo-search
python-dateutil
googlenewsdecoder==0.1.7
pandas
openpyxl
scikit-learn
cloudscraper
tqdm
